% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoSIMEX.R
\name{geoSIMEX}
\alias{geoSIMEX}
\title{Geographic SIMEX}
\usage{
geoSIMEX(model, geoSIMEXvariable, roiData, aidData, aid.project.amount, 
iterations=500, bins=3, fitting.method = "quadratic", roi.area="area",  
roi.pc1.name="pc1.id", roi.pc2.name="pc2.id", roi.pc3.name="pc3.id", 
roi.pc4.name="pc4.id", roi.pc5.name="pc5.id", roi.pc5.name="pc6.id",  
aid.pc1.centroid.name="centroid.pc1.id", aid.precision.code="precision.code",
parallel=TRUE, mc.cores=2)
}
\arguments{
\item{model}{the naive model}

\item{SIMEXvariable}{character containing the name of the variable with spatial uncertainty}

\item{roiData}{name of dataframe of ROI data}

\item{aidData}{name of dataframe of aid project data}

\item{aid.project.amount}{character containing the name of the variable in the aidData dataset which contains aid amounts (e.g., commitment, disbursement). Set value to 1 if interested in number of aid projects rather than dollars.}

\item{iterations}{number of simulated error iterations}

\item{bins}{number of bins to group coefficients}

\item{fitting.method}{fitting method for the extrapolation. linear and quadratic are implemented.}

\item{roi.area}{character containing the name of the variable in the ROI dataset which contains areas of ROIs. "area" is the default name in datasets produced by AidData's data extraction tool}

\item{roi.pc1.name}{character containing the name of the variable in the ROI dataset which contains names or IDs of the precision code 1 spatial area that each ROI falls within. "pc1.id" is the default name in datasets produced by AidData's data extraction tool}

\item{roi.pc2.name}{character containing the name of the variable in the ROI dataset which contains names or IDs of the precision code 2 spatial area that each ROI falls within. "pc2.id" is the default name in datasets produced by AidData's data extraction tool}

\item{roi.pc3.name}{character containing the name of the variable in the ROI dataset which contains names or IDs of the precision code 3 spatial area that each ROI falls within. "pc3.id" is the default name in datasets produced by AidData's data extraction tool}

\item{roi.pc4.name}{character containing the name of the variable in the ROI dataset which contains names or IDs of the precision code 4 spatial area that each ROI falls within. "pc4.id" is the default name in datasets produced by AidData's data extraction tool}

\item{roi.pc6.name}{character containing the name of the variable in the ROI dataset which contains names or IDs of the precision code 6 and 8 spatial area that each ROI falls within. "pc6.id" is the default name in datasets produced by AidData's data extraction tool}

\item{aid.pc1.centroid.name}{character containing the name of the variable in the aidData dataset which contains names or IDs of a precision code 1 spatial area that the aid project falls within. "centroid.pc1.id" is the default name in datasets produced by AidData's data extraction tool}

\item{aid.precision.code}{character containing the name of the variable in the aidData dataset which contains precision codes for each project. "pc1.id" is the default name in datasets produced by AidData's data extraction tool}

\item{parallel}{parallelize iterations}

\item{mc.cores}{number of cores to use for parallelization}
}
\description{
\code{geoSIMEX} Implementation of the geoSIMEX algorithm for models 
with spatial uncertainty. Package built to work with data from 
AidData's data extraction tool.
}
\details{
The values contained within roi.pc1.name and aid.pc1.centroid.name variables should be the same.
}
\note{
The function is built to work with data from AidData's data extration tool. The extraction tool can be accessed here: [provide website]. 

Just like the lm() and glm() packages, geoSIMEX() is equipped to work with stargazer.
}
\examples{
set.seed(42)

# Generating Country Dataset
countryData <- as.data.frame(matrix(NA, nrow=120, ncol=0))
countryData$pc1.id <- 1:120
countryData$pc2.id <- rep(1:(120/3), each=3)
countryData$pc3.id <- rep(1:(120/(3*2)), each=(3*2))
countryData$pc4.id <- rep(1:(120/(3*2*4)), each=(3*2*4))
countryData$pc6.id <- 1
countryData$area <- rgamma(120, shape=2)

# Creating Aid Dataset Without Error
aidData <- as.data.frame(matrix(NA,nrow=100,ncol=0))
aidData$disbursement <- runif(100,0,1) * 10000000 
aidData$centroid.pc1.id <- sample(size=100,x=c(1:120), prob=rep(1/120, 120), replace=TRUE)
aidData$precision.code <- 1

# Adding True Aid to Country Dataset
countryData$aid <- expected_aid_ROI(aidData=aidData, roiData=countryData, probAidAssume=countryData$area, aid.project.amount=aidData$disbursement, aid.precision.code="precision.code", roi.pc1.name="pc1.id", roi.pc2.name="pc2.id", roi.pc3.name="pc3.id", roi.pc4.name="pc4.id", roi.pc5.name="pc5.id", roi.pc6.name="pc6.id", aid.pc1.centroid.name="centroid.pc1.id")

# Defining True Relation Between Aid and Wealth
countryData$wealth <- countryData$aid + rnorm(120) * 0.1

# Creating Datasets with Uncertainty
countryData <- subset(countryData, select = -c(aid))
aidData$precision.code <- sample(size=100, x=c(1,2,3,4,6), prob=rep(1/5, 5), replace=TRUE)

# Calculating Expected Aid and Running Naive Model
countryData$Expected.Aid <- expected_aid_ROI(aidData=aidData, roiData=countryData, probAidAssume=countryData$area, aid.project.amount=aidData$disbursement, aid.precision.code="precision.code", roi.pc1.name="pc1.id", roi.pc2.name="pc2.id", roi.pc3.name="pc3.id", roi.pc4.name="pc4.id", roi.pc5.name="pc5.id", roi.pc6.name="pc6.id", aid.pc1.centroid.name="centroid.pc1.id")

lm_naive <- lm(wealth ~ Expected.Aid, data=countryData)

# Implementing GeoSIMEX
lm_geoSIMEX <- geoSIMEX(model = lm_naive, 
                        geoSIMEXvariable = "Expected.Aid", 
                        roiData = countryData, 
                        aidData = aidData, 
                        aid.project.amount = "disbursement")

summary(lm_geoSIMEX)
plot(lm_geoSIMEX, variable="Expected.Aid") 
}
\references{
Cook, J.R. and Stefanski, L.A. (1994) Simulation-extrapolation estimation in parametric measurement error models. Journal of American Statistical Association, 89, 1314 – 1328
}
\author{
AidData
}
